requirements.txt
selenium==4.18.1
pandas==2.2.1
openpyxl==3.1.2
webdriver-manager==4.0.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3 